LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-js-expose-expected.txt

Text

<<<<<<< HEAD
Harness Error (TIMEOUT), message = null

FAIL Element.getBoundingClientRect() returns the actual rendered location assert_equals: expected 200 but got 1500
FAIL Range.getBoundingClientRect() returns the actual rendered location assert_equals: expected 200 but got 1500
FAIL Element.offset* return adjusted offsets assert_equals: expected 200 but got 1500
TIMEOUT IntersectionObserver should fire when anchored element is moved into visible area Test timed out
=======
PASS Element.getBoundingClientRect() returns the actual rendered location
PASS Range.getBoundingClientRect() returns the actual rendered location
PASS Element.offset* return adjusted offsets
PASS IntersectionObserver should fire when anchored element is moved into visible area
>>>>>>> 90fc9aa7e180 (Make anchor positioning fallbacks respond to scrolling [css-anchor-position-1])
